pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-perotf"
description = "A schema package plugin for perotf at KIT."
version = "1.0.0"
readme = "README.rst"
requires-python = ">=3.10"
authors = [
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
    { name = "Felix Laufer", email = "felix.laufer@kit.edu" },

]
maintainers = [
    { name = "Felix Laufer", email = "felix.laufer@kit.edu" },
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.16",
    'nomad-baseclasses @ git+https://github.com/nomad-hzb/nomad-baseclasses.git',
    'nomad-measurements']

[project.urls]
Repository = "https://github.com/nomad-hzb/nomad-perotf"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint 
    "PL",
]

pylint.max-branches = 25
pylint.max-statements = 80

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLR2004",
    "PLR0913",
    "PLC0415"

] 

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than bs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
perotf_package = "nomad_perotf.schema_packages:perotf_package"
perotf_parser = "nomad_perotf.parsers:perotf_parser"
perotf_experiment_parser = "nomad_perotf.parsers:perotf_experiment_parser"

#voila app
perotf_voila_app = "nomad_perotf.apps:perotf_voila_app"
solar_cell_overview_entry_point = "nomad_perotf.apps:solar_cell_overview_entry_point"

